+++ smol/cofe/space/clusterspace.py
@@ -11,6 +11,7 @@ from copy import deepcopy
 from importlib import import_module
 import warnings
 import numpy as np
+from itertools import combinations
 
 from monty.json import MSONable
 from pymatgen import Structure, PeriodicSite
@@ -29,7 +30,8 @@ from smol.cofe.space.constants import SITE_TOL
 from smol.exceptions import \
     SymmetryError, StructureMatchError, SYMMETRY_ERROR_MESSAGE
 
-__author__ = "Luis Barroso-Luque, William Davidson Richard"
+__author__ = "Luis Barroso-Luque, William Davidson Richards, Fengyu Xie," \
+             "Peichen Zhong"
 
 
 class ClusterSubspace(MSONable):
@@ -93,12 +95,12 @@ class ClusterSubspace(MSONable):
             orbits (dict): {size: list of Orbits}
                 Dictionary with size (number of sites) as keys and list of
                 Orbits as values.
-            supercell_matcher (StructureMatcher): (optional)
+            supercell_matcher (StructureMatcher): optional
                 A StructureMatcher class to be used to find supercell matrices
                 relating the prim structure to other structures. If you pass
                 this directly you should know how to set the matcher up, other
                 wise matching your relaxed structures can fail, alot.
-            site_matcher (StructureMatcher): (optional)
+            site_matcher (StructureMatcher): optional
                 A StructureMatcher class to be used to find site mappings
                 relating the sites of a given structure to an appropriate
                 supercell of the prim structure . If you pass this directly you
@@ -165,6 +167,10 @@ class ClusterSubspace(MSONable):
         # already been matched
         self._supercell_orb_inds = {}
 
+        # 2D lists to store 1-level-down hierarchy info. (One level only!)
+        # Will be cleaned after any change of orbits!
+        self._bit_combo_hierarchy = None
+
         # assign the cluster ids
         self._assign_orbit_ids()
 
@@ -206,12 +212,12 @@ class ClusterSubspace(MSONable):
                 If true the concentrations in the prim structure sites will be
                 used to orthormalize site bases. This gives gives a cluster
                 subspace centered about the prim composition.
-            supercell_matcher (StructureMatcher): (optional)
+            supercell_matcher (StructureMatcher): optional
                 A StructureMatcher class to be used to find supercell matrices
                 relating the prim structure to other structures. If you pass
                 this directly you should know how to set the matcher up other
                 wise matching your relaxed structures will fail, alot.
-            site_matcher (StructureMatcher): (optional)
+            site_matcher (StructureMatcher): optional
                 A StructureMatcher class to be used to find site mappings
                 relating the sites of a given structure to an appropriate
                 supercell of the prim structure . If you pass this directly you
@@ -293,6 +299,16 @@ class ClusterSubspace(MSONable):
         return np.array(mults)
 
     @property
+    def all_bit_combos(self):
+        """Return flattened all bit_combos for each correlation function.
+
+        A list of all bit_combos of length self.num_corr_functions.
+        This allows to obtain a bit combo by bit id (empty cluster has None)
+        """
+        return [None] + [combos for orbit in self.orbits
+                         for combos in orbit.bit_combos]
+
+    @property
     def num_functions_per_orbit(self):
         """Get the number of correlation functions for each orbit.
 
@@ -344,6 +360,53 @@ class ClusterSubspace(MSONable):
         """
         return self.orbit_multiplicities[self.function_orbit_ids] * \
             self.function_ordering_multiplicities
+
+    def bit_combo_hierarchy(self, min_size=2, invert=False):
+        """Get 1-level-down hierarchy of correlation functions.
+
+        The size difference between the current corr function and its
+        sub-clusters is only 1! We only give one-level down hierarchy. Because
+        that is enough to contrain hierarchy!
+
+        Note: Since complete, all level hierarchy table is not practical
+              for CE fit, we will not include it as an attribute of this class.
+              If you still want to see it, call function get_complete_mapping
+              in this module.
+
+        Args:
+            min_size (int): optional
+                Minimum size required for the correlation function. If the size
+                of the correlation function is smaller or equals to min_size,
+                will not search for its sub-clusters. For hierarchy
+                constraints, the recommended setting is 2.
+            invert (bool): optional
+                Default is invert=False which gives the high to low bit combo
+                hierarchy. Invert= True will invert the hierarchy into low to
+                high
+
+        Returns:
+            list of lists: Each sublist is of of length self.num_corr_function
+            and contains integer indices of correlation functions that are
+            contained by the correlation function with the current index.
+        """
+        if self._bit_combo_hierarchy is None:
+            self._bit_combo_hierarchy = self._get_hierarchy_up_to_low()
+
+        # array of orbit sizes for each bit id.
+        all_sizes = np.array([0] + [self.orbits[i - 1].base_cluster.size
+                                    for i in self.function_orbit_ids[1:]])
+
+        # Stores a hierarchy from cluster size 1, then retrieves from min_size.
+        hierarchy = []
+        for vals, size in zip(self._bit_combo_hierarchy, all_sizes):
+            if size <= min_size:
+                hierarchy.append([])
+            else:
+                hierarchy.append(vals)
+        if invert:
+            return invert_mapping_table(hierarchy)
+        else:
+            return hierarchy
 
     @property
     def basis_orthogonal(self):
@@ -690,10 +753,11 @@ class ClusterSubspace(MSONable):
                                   if orbit.id not in orbit_ids]
 
         self._assign_orbit_ids()  # Re-assign ids
-        # Clear the cached supercell orbit mappings
+        # Clear the cached supercell orbit mappings and hierarchy
         self._supercell_orb_inds = {}
+        self._bit_combo_hierarchy = None
 
-    def remove_orbit_bit_combos(self, orbit_bit_ids):
+    def remove_orbit_bit_combos(self, bit_ids):
         """Remove orbit bit combos by their ids.
 
         Removes a specific bit combo from an orbit. This allows more granular
@@ -711,11 +775,11 @@ class ClusterSubspace(MSONable):
         does anyway...
 
         Args:
-            orbit_bit_ids (list):
+            bit_ids (list):
                 list of orbit bit ids to remove
         """
         empty_orbit_ids = []
-        bit_ids = np.array(orbit_bit_ids, dtype=int)
+        bit_ids = np.array(bit_ids, dtype=int)
 
         for orbit in self.iterorbits():
             first_id = orbit.bit_id
@@ -735,6 +799,9 @@ class ClusterSubspace(MSONable):
             self.remove_orbits(empty_orbit_ids)
         else:
             self._assign_orbit_ids()  # Re-assign ids
+
+        # clear hierarchy
+        self._bit_combo_hierarchy = None
 
     def copy(self):
         """Deep copy of instance."""
@@ -897,6 +964,92 @@ class ClusterSubspace(MSONable):
 
         return orbit_indices
 
+    def _find_sub_cluster(self, bit_id, min_size=1):
+        """Find 1-level-down subclusters of a given correlation function.
+
+        Args:
+            bit_id (int):
+                Index of the correlation function to find subclusters with.
+            min_size (int): optional
+                Minimum size required for the correlation function. If  the
+                size of the correlation function is smaller or equals to
+                min_size, will not search for its sub-clusters.
+
+        Returns:
+            list: A list of integer indices specifying which correlation
+            functions are 1-level-down subclusters of the given correlation
+            function.
+        """
+        if bit_id == 0:  # Constant term
+            return []
+
+        # Separate the zero term out.
+        all_sizes = np.array([0] + [self.orbits[i - 1].base_cluster.size
+                             for i in self.function_orbit_ids[1:]])
+
+        bit_combos = self.all_bit_combos[bit_id]
+        orbit = self.orbits[self.function_orbit_ids[bit_id] - 1]
+        sites = orbit.base_cluster.sites
+
+        size = all_sizes[bit_id]
+        if size <= min_size:
+            return []
+
+        possible_sub_ids = np.where(all_sizes == (size - 1))[0]
+        lattice = self._exp_structure.lattice
+        sub_indices = []
+        for comb in combinations(np.arange(size), size - 1):
+            sub_sites = np.array(sites[np.array(comb), :])
+            sub_bit_combos = np.array(bit_combos[:, np.array(comb)])
+            sub_cluster = Cluster(sub_sites, lattice)
+            sub_equiv = [sub_cluster]
+            for symop in self.symops:
+                new_sites = symop.operate_multi(sub_sites)
+                c = Cluster(new_sites, lattice)
+                if c not in sub_equiv:
+                    sub_equiv.append(c)
+
+            for sub_id in possible_sub_ids:
+                sub_bit_combo = self.all_bit_combos[sub_id]
+                sub_orbit = self.orbits[self.function_orbit_ids[sub_id] - 1]
+                cluster_match = sub_orbit.base_cluster in sub_equiv
+                bit_match = np.any(
+                    np.all(sub_bit_combo[0] == sub_bit_combos, axis=1))
+
+                if cluster_match and bit_match:
+                    if sub_id in sub_indices:
+                        continue
+                    else:
+                        sub_indices.append(sub_id)
+
+        return sub_indices
+
+    def _get_hierarchy_up_to_low(self, min_size=1):
+        """Generate high-to-low hierarchy.
+
+        The size difference between the current corr function and its
+        sub clusters is only 1! We only give one-level down hierarchy
+        Because it would be enough to contrain hierarchy!
+
+        Args:
+            min_size:
+                Minimum size required for the correlation function. If  the
+                size of the correlation function is smaller or equals to
+                min_size, will not search for its sub-clusters.
+
+        Returns:
+            list of lists: Each sublist of length self.num_corr_function
+            contains integer indices of correlation functions that are
+            contained by the correlation function with the current index.
+        """
+        up_low_hierarchy = [[] for i in range(self.num_corr_functions)]
+
+        for ii in np.flip(np.arange(self.num_corr_functions)):
+            sub_indices = self._find_sub_cluster(ii, min_size=min_size)
+            up_low_hierarchy[ii] = sub_indices
+
+        return up_low_hierarchy
+
     def __eq__(self, other):
         """Check equality between cluster subspaces."""
         if not isinstance(other, ClusterSubspace):
@@ -970,6 +1123,7 @@ class ClusterSubspace(MSONable):
                                         np.array(ind)) for o_id, ind
                                         in orb_inds]
         cs._supercell_orb_inds = _supercell_orb_inds
+        cs._bit_combo_hierarchy = d.get('_bc_hierarchy')
         return cs
 
     def as_dict(self):
@@ -993,8 +1147,73 @@ class ClusterSubspace(MSONable):
              'sc_matcher': self._sc_matcher.as_dict(),
              'site_matcher': self._site_matcher.as_dict(),
              'external_terms': [et.as_dict() for et in self.external_terms],
-             '_supercell_orb_inds': _supercell_orb_inds}
+             '_supercell_orb_inds': _supercell_orb_inds,
+             '_bc_hierarchy': self._bit_combo_hierarchy}
         return d
+
+
+def invert_mapping_table(mapping):
+    """Invert a mapping table from forward to backward, vice versa.
+
+    Args:
+        mapping (list of lists):
+            List of sublists, each contains integer indices, indicating
+            a foward mapping from the current sublist index to the indices
+            in the sublist.
+
+    Returns:
+        lists of lists: Inverted mapping table containing backward mapping.
+        Same format as input.
+    """
+    inv_mapping = [[] for _ in range(len(mapping))]
+
+    for i in range(len(mapping) - 1, -1, -1):
+        values_list = mapping[i]
+        for value in values_list:
+            if not (i in inv_mapping[value]):
+                inv_mapping[value].append(i)
+
+    return inv_mapping
+
+
+def get_complete_mapping(mapping):
+    """Get a complete mapping from a 1 level mapping.
+
+    If we allow transferability between mapping linkages, there would be
+    a full mapping containing all linkages at all connectivity level.
+    Using this function, you can get a full mapping from an incomplete
+    mapping.
+
+    NOTE: Since complete hierarchy is not very useful for actual CE fit, we
+    will not include it as an attribute of ClusterSubspace.
+
+    Args:
+        mapping (list of lists):
+             List of sublists, each contains integer indices, indicating
+            a foward mapping from the current sublist index to the indices
+            in the sublist.
+
+    Returns:
+        list of lists: Full mapping table containing forward mapping, but with
+        all connectivity levels. Same format as input
+    """
+    all_level_mapping = deepcopy(mapping)
+
+    for i in range(len(mapping) - 1, -1, -1):
+        next_values_list = mapping[i]
+
+        while len(next_values_list) > 0:
+            for next_value in next_values_list:
+                if next_value not in all_level_mapping[i]:
+                    all_level_mapping[i].append(next_value)
+            next_values_list_new = []
+            for next_value in next_values_list:
+                for nn_value in mapping[next_value]:
+                    if nn_value not in next_values_list_new:
+                        next_values_list_new.append(nn_value)
+            next_values_list = next_values_list_new
+
+    return all_level_mapping
 
 
 class PottsSubspace(ClusterSubspace):